Put Therapy Setting descriptive text in a common place
