+++ LoopKitUI/Views/Information Screens/CorrectionRangeOverrideInformationView.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2020 LoopKit Authors. All rights reserved.
 //
 
+import LoopKit
 import SwiftUI
 
 public struct CorrectionRangeOverrideInformationView: View {
