+++ LoopKitUI/Views/Settings Editors/CorrectionRangeOverridesEditor.swift
@@ -10,23 +10,21 @@ import SwiftUI
 import HealthKit
 import LoopKit
 
-
-public struct CorrectionRangeOverrides: Equatable {
-    enum Preset: Hashable, CaseIterable {
-        case preMeal
-        case workout
+extension TherapySetting {
+    
+    public func helpScreen() -> some View {
+        switch self {
+        case .glucoseTargetRange:
+            return AnyView(CorrectionRangeInformationView(onExit: nil, mode: .modal))
+        case .correctionRangeOverrides:
+            return AnyView(CorrectionRangeOverrideInformationView(onExit: nil, mode: .modal))
+        case .suspendThreshold:
+            return AnyView(SuspendThresholdInformationView(onExit: nil, mode: .modal))
+        // ANNA TODO: add more once other instructional screens are created
+        default:
+            return AnyView(Text("To be implemented"))
+        }
     }
-
-    var ranges: [Preset: ClosedRange<HKQuantity>]
-
-    public init(preMeal: DoubleRange?, workout: DoubleRange?, unit: HKUnit) {
-        ranges = [:]
-        ranges[.preMeal] = preMeal?.quantityRange(for: unit)
-        ranges[.workout] = workout?.quantityRange(for: unit)
-    }
-
-    public var preMeal: ClosedRange<HKQuantity>? { ranges[.preMeal] }
-    public var workout: ClosedRange<HKQuantity>? { ranges[.workout] }
 }
 
 public struct CorrectionRangeOverridesEditor: View {
@@ -138,9 +136,9 @@ public struct CorrectionRangeOverridesEditor: View {
     private func description(of preset: CorrectionRangeOverrides.Preset) -> Text {
         switch preset {
         case .preMeal:
-            return Text("Temporarily lower your glucose target before a meal to impact post-meal glucose spikes.", comment: "Description of pre-meal mode")
+            return Text(preset.descriptiveText)
         case .workout:
-            return Text("Temporarily raise your glucose target before, during, or after physical activity to reduce the risk of low glucose events.", comment: "Description of workout mode")
+            return Text(preset.descriptiveText)
         }
     }
